Re-jiggered a bunch of code because I had mangled some ParSeq things.
The short version is that I was treating a comprehension over a ParSeq
like a comprehension over a Stream, and assuming that results would be
lazy when they weren't.  This was causing all sorts of problems
because it was building up an entire list of results rather than
streaming them to a file as intended.  This was preventing results
from being written incrementally (as intended), and it was also the
source of a crash on a large test.